requirements.txt
# 核心框架
torch>=2.1.0
transformers>=4.37.0
# 导出工具（onnxruntime-gpu 同时支持 -d cpu 与 -d cuda）
optimum>=1.17.0
onnxruntime-gpu
accelerate  # ONNX 导出时权重去重检查需要
# 图优化工具
onnx>=1.15.0
onnxslim>=0.1.10
# 下载增强
huggingface_hub
modelscope
